refactor(core): cache commands/nodeViews, extract spec helpers, split builtIn.ts into domain files; fix bubble menu in table cells
