+++ packages/core/package.json
@@ -48,7 +48,6 @@
     "prosemirror-model": "^1.25.4",
     "prosemirror-schema-list": "^1.5.1",
     "prosemirror-state": "^1.4.4",
-    "prosemirror-tables": "^1.8.5",
     "prosemirror-transform": "^1.10.5",
     "prosemirror-view": "^1.41.5"
   },
